# Enable `re_renderer` for all tests using the `TestContext` if graphics adapter is present