--- a/crates/viewer/re_component_ui/tests/snapshots/all_components_list_item_narrow/Colormap_placeholder.png
+++ b/crates/viewer/re_component_ui/tests/snapshots/all_components_list_item_narrow/Colormap_placeholder.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:18b8ada626cbb59ec85db0ba8d167deeef516d18279d5b85ad3dd43bef7b8113
-size 3326
+oid sha256:addb336fa14268b87e22974902909b3a3e5629fbe483f7d98fe9ea04f2836e47
+size 2986