--- a/crates/viewer/re_component_ui/tests/snapshots/all_components_list_item_wide/Colormap_placeholder.png
+++ b/crates/viewer/re_component_ui/tests/snapshots/all_components_list_item_wide/Colormap_placeholder.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:b15bc121e4102dbc979e92f5564c36ad7f085a3c9465518f3d90427d0fbe3f38
-size 3637
+oid sha256:64a67110962fafff3f8bef5f849b5401a6d7f2adebf3a3ac82d286ef70eea7ed
+size 4208