--- a/crates/viewer/re_viewer_context/Cargo.toml
+++ b/crates/viewer/re_viewer_context/Cargo.toml
@@ -20,7 +20,7 @@ all-features = true
 
 [features]
 ## Enable for testing utilities.
-testing = []
+testing = ["dep:pollster", "dep:egui_kittest", "dep:env_logger"]
 
 [dependencies]
 re_capabilities.workspace = true
@@ -72,6 +72,11 @@ strum_macros.workspace = true
 thiserror.workspace = true
 uuid = { workspace = true, features = ["serde", "v4", "js"] }
 wgpu.workspace = true
+
+# Optional dependencies:
+egui_kittest = { workspace = true, features = ["wgpu"], optional = true }
+env_logger = { workspace = true, optional = true }
+pollster = { workspace = true, optional = true }
 
 # Native dependencies:
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]