The viewer test harness now tries to create a wgpu adapter (only once via a `Lazy`) in order to re-use this in an re_renderer context. As a consequence screenshot tests can now use egui elements that bottom out in `re_renderer` drawing (i.e. in `ViewBuilder`).

For the moment the adapter creation is allowed to fail, meaning that if we're on a ci machine without any graphics adapter at all, we can still continue as before with render context == `None`.

* [x] pass main ci
* [x] What's the performance before/after for component tests?

## Performance:

Timing reports for `cargo test -p re_component_ui` (as reported by test runner) on Windows:

* on this branch
  * first run after checkout: 2.34s
  * next two successive rounds: 2.30s, 2.12s
* on main with snapshots enabled (commenting out the still remaining mac-only guard):
  * first run after checkout: 31.86s
  * next two successive rounds: 26.70s, 28.61s
* on main without snapshots:
  * first run after checkout: 0.43s
  * next two successive rounds: 0.42s, 0.44s


## Next steps:
* enable this on for Linux & Windows CI -> copy wgpu's software rasterizer setup
    * make it an error when no adapter can be created?
* make `render_ctx` non option again
* improve setup using new abilities from https://github.com/emilk/egui/pull/5506