@Wumpf 2025-01-07
note to self: expecting this to fail since the PR as is requires graphics adapter creation to be always successful and we don't do anything towards that. But I want to see how it fails exactly and then remedy it