@Wumpf 2025-01-08
decided to take this more step by step: for the moment this does not enforce the wgpu adapter and by extension the re_renderer context to be present