@Wumpf 2025-01-09
> But I think it would be great if we got a panic if we forget to call setup_kittest_for_rendering when re_renderer is needed

good idea, I look into it!