@Wumpf 2025-01-09
Tried a bit and it looks doable.. **if** re_renderer is always created. Luckily the next PR takes care of that anyways!